ai-web-crawler/pyproject.toml
[project]
name = "ai-web-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.7.4",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
]
